requirements.txt
streamlit
openai>=1.0.0
boto3
python-dotenv
scikit-learn

